# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MirrorTech POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions
      security:
        - bearerAuth: []
      parameters:
        - name: locationId
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: paymentMethod
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a new sale transaction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a single transaction
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}/void:
    post:
      operationId: voidTransaction
      tags: [transactions]
      summary: Void a transaction
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoidInput"
      responses:
        "200":
          description: Voided
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}/receipt:
    get:
      operationId: getReceipt
      tags: [transactions]
      summary: Get receipt data for a transaction
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Receipt data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receipt"
  /transactions/sync:
    post:
      operationId: syncOfflineTransactions
      tags: [transactions]
      summary: Sync offline transactions from localStorage queue
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncInput"
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResult"
...
  /transfers:
    get:
      operationId: listTransfers
      tags: [transfers]
      summary: List stock transfers
      security:
        - bearerAuth: []
      parameters:
        - name: locationId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of transfers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transfer"
    post:
      operationId: createTransfer
      tags: [transfers]
      summary: Request a stock transfer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
  /transfers/{id}/approve:
    post:
      operationId: approveTransfer
      tags: [transfers]
      summary: Approve a transfer and move stock
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Location:
      type: object
      required: [id, name, isActive]
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
...
    CartItem:
      type: object
      required: [itemId, name, quantity, price]
      properties:
        itemId:
          type: string
        name:
          type: string
        quantity:
          type: integer
        price:
          type: string
        total:
          type: string
    TransactionInput:
      type: object
      required: [locationId, items, subtotal, taxAmount, total, paymentMethod]
      properties:
        locationId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: string
        taxAmount:
          type: string
        total:
          type: string
        paymentMethod:
          type: string
          enum: [cash, momo, card]
        momoPhone:
          type: string
        momoNetwork:
          type: string
        momoReference:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        notes:
          type: string
    Transaction:
      type: object
      required: [id, receiptNumber, locationId, items, subtotal, taxAmount, total, paymentMethod, paymentStatus, createdAt]
      properties:
        id:
          type: string
        receiptNumber:
          type: string
        locationId:
          type: string
        locationName:
          type: ["string", "null"]
        cashierId:
          type: ["string", "null"]
        cashierName:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: string
        taxAmount:
          type: string
        total:
          type: string
        paymentMethod:
          type: string
        paymentStatus:
          type: string
        momoPhone:
          type: ["string", "null"]
        momoNetwork:
          type: ["string", "null"]
        momoReference:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isVoided:
          type: boolean
        voidReason:
          type: ["string", "null"]
        createdAt:
          type: string
    TransactionList:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
    VoidInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
    Receipt:
      type: object
      required: [transaction, location]
      properties:
        transaction:
          $ref: "#/components/schemas/Transaction"
        location:
          $ref: "#/components/schemas/Location"
        graReceiptNumber:
          type: string
        qrCode:
          type: ["string", "null"]
    SyncInput:
      type: object
      required: [transactions]
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/TransactionInput"
    SyncResult:
      type: object
      required: [synced, failed]
      properties:
        synced:
          type: integer
        failed:
          type: integer
        errors:
          type: array
          items:
            type: string
...
    TransferInput:
      type: object
      required: [itemId, fromLocationId, toLocationId, quantity]
      properties:
        itemId:
          type: string
        fromLocationId:
          type: string
        toLocationId:
          type: string
        quantity:
          type: integer
        notes:
          type: string
    Transfer:
      type: object
      required: [id, itemId, fromLocationId, toLocationId, quantity, status]
      properties:
        id:
          type: string
        itemId:
          type: string
        itemName:
          type: ["string", "null"]
        fromLocationId:
          type: string
        fromLocationName:
          type: ["string", "null"]
        toLocationId:
          type: string
        toLocationName:
          type: ["string", "null"]
        quantity:
          type: integer
        status:
          type: string
        notes:
          type: ["string", "null"]
        requestedById:
          type: ["string", "null"]
        approvedById:
          type: ["string", "null"]
        createdAt:
          type: string
...