# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MirrorTech POS API
 * OpenAPI spec version: 0.1.0
 */

export interface CartItem {
  itemId: string;
  name: string;
  quantity: number;
  price: string;
  total?: string;
}